backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
bcrypt==4.1.3
aiofiles==23.2.1
Pillow>=10.0.0
